research_documentation/alert_django/requirements.txt
# DJANGO RELATED
django-countries>=4.6.1
django>=1.11.5
django-filter>=1.0.1
pytz==2017.2
six==1.11.0
mysqlclient==1.3.12
djangorestframework==3.7.3
markdown>=2.1.0

# Data import
django-import-export==1.0.0

# CORE ASTRO
astropy>=2.0.2
numpy==1.13.1
pyvo==0.6.1
astroquery==0.3.5
pandas
scipy>=0.17.0

# PLOTTING
matplotlib==2.2.2
#vaex>=1.0.0b8
#ipyvolume
bokeh==0.12.15
Pillow==4.3.0
matplotlib==2.2.2
APLpy>=2.0rc2

# MULTI-THREADING
celery==4.2
redis

#For dev:
django-extensions==1.9.7
ipython==5.6.0
future
